+++ common/src/main/java/club/iananderson/seasonhud/impl/minimaps/CurrentMinimap.java
@@ -5,15 +5,10 @@ import club.iananderson.seasonhud.impl.seasons.Calendar;
 import club.iananderson.seasonhud.platform.Services;
 import java.util.ArrayList;
 import java.util.List;
-import journeymap.client.ui.UIManager;
-import journeymap.client.ui.fullscreen.Fullscreen;
 import journeymap.client.ui.minimap.MiniMap;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.screens.ChatScreen;
 import net.minecraft.client.gui.screens.DeathScreen;
-import net.minecraft.client.gui.screens.Screen;
-import net.minecraft.client.gui.screens.inventory.CreativeModeInventoryScreen;
-import net.minecraft.client.gui.screens.inventory.InventoryScreen;
 
 public class CurrentMinimap {
   private static boolean minimapLoaded(Minimaps minimap) {
@@ -71,13 +66,13 @@ public class CurrentMinimap {
     if (journeyMapLoaded()) {
       validScreens = MiniMap.uiState().active;
     }
+    if (ftbChunksLoaded()) {
+      validScreens = true;
+    }
 
     boolean hiddenMinimap = Services.MINIMAP.hideHudInCurrentDimension() || Services.MINIMAP.hiddenMinimap(minimap);
 
     return enabled && Calendar.calendarFound() && validScreens && !mc.options.hideGui && !hiddenMinimap;
-    return Config.getEnableMod() && Config.getEnableMinimapIntegration() && Calendar.calendarFound() && validScreen(
-        mc.screen) && !mc.options.hideGui && !Services.MINIMAP.hideHudInCurrentDimension()
-        && !Services.MINIMAP.hiddenMinimap(minimap);
   }
 
   public enum Minimaps {