+++ common/src/main/java/club/iananderson/seasonhud/impl/seasons/Calendar.java
@@ -3,6 +3,8 @@ package club.iananderson.seasonhud.impl.seasons;
 import club.iananderson.seasonhud.Common;
 import club.iananderson.seasonhud.config.Config;
 import club.iananderson.seasonhud.platform.Services;
+import io.wispforest.accessories.api.AccessoriesCapability;
+import io.wispforest.accessories.api.AccessoriesContainer;
 import net.minecraft.client.Minecraft;
 import net.minecraft.world.entity.player.Inventory;
 import net.minecraft.world.item.Item;
@@ -19,10 +21,19 @@ public class Calendar {
       return false;
     }
 
-    return findCalendar(mc.player.getInventory(), calendar) || !Config.getNeedCalendar();
+    boolean curio = false;
+
+    if (Common.accessoriesLoaded() && !Common.curiosLoaded()) {
+      AccessoriesContainer accessoriesContainer = AccessoriesCapability.get(mc.player).getContainers()
+          .get("calendarslot");
+      curio = accessoriesContainer.getAccessories().countItem(calendar) > 0;
+    }
+
+    return findCalendar(mc.player.getInventory(), calendar) || curio || !Config.getNeedCalendar();
   }
 
   private static boolean findCalendar(Inventory inv, Item item) {
+
     return inv.contains(item.getDefaultInstance());
   }
 }