+++ fabric/src/main/java/club/iananderson/seasonhud/fabric/client/SeasonHudClientFabric.java
@@ -3,6 +3,9 @@ package club.iananderson.seasonhud.fabric.client;
 import club.iananderson.seasonhud.Common;
 import club.iananderson.seasonhud.config.Config;
 import club.iananderson.seasonhud.fabric.event.ClientEvents;
+import club.iananderson.seasonhud.impl.minimaps.CurrentMinimap.Minimaps;
+import club.iananderson.seasonhud.impl.minimaps.SeasonComponent;
+import club.iananderson.seasonhud.platform.Services;
 import fuzs.forgeconfigapiport.fabric.api.neoforge.v4.NeoForgeConfigRegistry;
 import net.fabricmc.api.ClientModInitializer;
 import net.neoforged.fml.config.ModConfig;
@@ -14,5 +17,9 @@ public class SeasonHudClientFabric implements ClientModInitializer {
     NeoForgeConfigRegistry.INSTANCE.register(Common.MOD_ID, ModConfig.Type.CLIENT, Config.GENERAL_SPEC,
                                              "SeasonHUD-client.toml");
     ClientEvents.register();
+
+    if (Services.PLATFORM.isModLoaded(Minimaps.FTB_CHUNKS.getModID())) {
+      SeasonComponent.registerFtbSeason();
+    }
   }
 }