+++ fabric/src/main/java/club/iananderson/seasonhud/fabric/impl/ModMenuImpl.java
@@ -1,4 +1,0 @@
-package club.iananderson.seasonhud.fabric.impl;
-
-public class ModMenuImpl {
-}