+++ neoforge/src/main/java/club/iananderson/seasonhud/neoforge/client/overlays/SeasonHUDOverlay.java
@@ -1,10 +1,9 @@
-package club.iananderson.seasonhud.forge.client.overlays;
+package club.iananderson.seasonhud.neoforge.client.overlays;
 
 import club.iananderson.seasonhud.client.overlays.SeasonHUDOverlayCommon;
 import net.minecraft.client.DeltaTracker;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.gui.LayeredDraw;
-import org.jetbrains.annotations.NotNull;
 
 public class SeasonHUDOverlay implements LayeredDraw.Layer {
   public static SeasonHUDOverlay HUD_INSTANCE;
@@ -13,8 +12,7 @@ public class SeasonHUDOverlay implements LayeredDraw.Layer {
     HUD_INSTANCE = new SeasonHUDOverlay();
   }
 
-  @Override
-  public void render(@NotNull GuiGraphics graphics, @NotNull DeltaTracker deltaTracker) {
-    SeasonHUDOverlayCommon.render(graphics);
+  public void render(GuiGraphics seasonStack, DeltaTracker deltaTracker) {
+    SeasonHUDOverlayCommon.render(seasonStack);
   }
 }